Search response

+++ src/main/java/modules/commons/search/request/Aggregation.java
@@ -3,18 +3,19 @@ package modules.commons.search.request;
 public class Aggregation {
 	
 	private String name;
-	private int count;
+	private Long count;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getCount() {
+	public Long getCount() {
 		return count;
 	}
-	public void setCount(int count) {
+	public void setCount(Long count) {
 		this.count = count;
 	}
+
 
 }
